separate LatestInstaller and SnapshotInstaller
